Login using challenge instead of encrypted wif to server
